# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Charge Watcher
  version: ""
host: localhost:8088
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
  /devices/{deviceID}/charge:
    post:
      tags:
      - chargewatch
      summary: updateCharge chargewatch
      operationId: chargewatch#updateCharge
      parameters:
      - name: deviceID
        in: path
        description: deviceID
        required: true
        type: string
      - name: object
        in: body
        required: true
        schema:
          type: object
          properties:
            chargeValue:
              type: integer
              example: 8894378129742806907
              format: int64
            charging:
              type: boolean
              example: true
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ChargewatchUpdateChargeResponseBody'
            required:
            - device
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/ChargewatchUpdateChargeStatusBadRequestResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/ChargewatchUpdateChargeStatusInternalServerErrorResponseBody'
      schemes:
      - http
  /devices/{deviceID}/chargeHistory:
    get:
      tags:
      - chargewatch
      summary: getChargeHistory chargewatch
      operationId: chargewatch#getChargeHistory
      parameters:
      - name: deviceID
        in: path
        description: deviceID
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ChargewatchGetChargeHistoryResponseBody'
            required:
            - chargeHistory
      schemes:
      - http
...
  /user/{userID}/devices:
    get:
      tags:
      - chargewatch
      summary: listDevices chargewatch
      operationId: chargewatch#listDevices
      parameters:
      - name: userID
        in: path
        description: userID
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ChargewatchListDevicesResponseBody'
            required:
            - devices
      schemes:
      - http
    post:
      tags:
      - chargewatch
      summary: createDevice chargewatch
      operationId: chargewatch#createDevice
      parameters:
      - name: userID
        in: path
        description: userID
        required: true
        type: string
      - name: object
        in: body
        required: true
        schema:
          type: object
          properties:
            description:
              type: string
              description: description
              default: ""
              example: Ducimus expedita iste dolore commodi.
            name:
              type: string
              description: device name
              example: Esse iste sunt sed et expedita.
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ChargewatchCreateDeviceResponseBody'
            required:
            - device
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/ChargewatchCreateDeviceStatusInternalServerErrorResponseBody'
      schemes:
      - http
  /user/{userID}/devices/{deviceID}:
    post:
      tags:
      - chargewatch
      summary: updateDevice chargewatch
      operationId: chargewatch#updateDevice
      parameters:
      - name: userID
        in: path
        description: userID
        required: true
        type: string
      - name: deviceID
        in: path
        description: deviceID
        required: true
        type: string
      - name: object
        in: body
        required: true
        schema:
          type: object
          properties:
            chargeValue:
              type: integer
              description: value [0,100]
              example: 11
              minimum: 0
              maximum: 100
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ChargewatchUpdateDeviceResponseBody'
            required:
            - device
      schemes:
      - http
definitions:
  ChargeHistoryResponseBody:
    title: ChargeHistoryResponseBody
    type: object
    properties:
      deviceID:
        type: string
        example: Dignissimos commodi odit.
      logs:
        type: array
        items:
          $ref: '#/definitions/ChargeLogResponseBody'
        example:
        - charge:
            charging: true
            timestamp: 7530045597555114181
            value: 7362310245545543380
          deviceID: Quia provident rerum.
        - charge:
            charging: true
            timestamp: 7530045597555114181
            value: 7362310245545543380
          deviceID: Quia provident rerum.
        - charge:
            charging: true
            timestamp: 7530045597555114181
            value: 7362310245545543380
          deviceID: Quia provident rerum.
        - charge:
            charging: true
            timestamp: 7530045597555114181
            value: 7362310245545543380
          deviceID: Quia provident rerum.
    example:
      deviceID: Est et alias assumenda et voluptatem qui.
      logs:
      - charge:
          charging: true
          timestamp: 7530045597555114181
          value: 7362310245545543380
        deviceID: Quia provident rerum.
      - charge:
          charging: true
          timestamp: 7530045597555114181
          value: 7362310245545543380
        deviceID: Quia provident rerum.
      - charge:
          charging: true
          timestamp: 7530045597555114181
          value: 7362310245545543380
        deviceID: Quia provident rerum.
      - charge:
          charging: true
          timestamp: 7530045597555114181
          value: 7362310245545543380
        deviceID: Quia provident rerum.
    required:
    - deviceID
    - logs
  ChargeLogResponseBody:
    title: ChargeLogResponseBody
    type: object
    properties:
      charge:
        $ref: '#/definitions/ChargeResponseBody'
      deviceID:
        type: string
        example: Est nesciunt impedit non aspernatur.
    example:
      charge:
        charging: true
        timestamp: 7530045597555114181
        value: 7362310245545543380
      deviceID: Provident consequatur beatae in eveniet et dignissimos.
    required:
    - deviceID
    - charge
  ChargeResponseBody:
    title: ChargeResponseBody
    type: object
    properties:
      charging:
        type: boolean
        example: true
      timestamp:
        type: integer
        example: 5757041125096478063
        format: int64
      value:
        type: integer
        example: 6073548835115809330
        format: int64
    example:
      charging: true
      timestamp: 3668634551682271274
      value: 8600777957003406859
    required:
    - value
    - charging
    - timestamp
  ChargewatchCreateDeviceResponseBody:
    title: ChargewatchCreateDeviceResponseBody
    type: object
    properties:
      device:
        $ref: '#/definitions/DeviceResponseBody'
    example:
      device:
        charge:
          charging: false
          timestamp: 8344681023334431344
          value: 1969285450001603711
        description: Sit ea earum vel.
        id: Iusto optio.
        name: Quia quas illo sunt quae vitae.
        state: Quaerat totam.
        userID: Corporis labore accusamus sit excepturi non.
    required:
    - device
  ChargewatchCreateDeviceStatusInternalServerErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: createDevice_StatusInternalServerError_response_body result type
      (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ChargewatchGetChargeHistoryResponseBody:
    title: ChargewatchGetChargeHistoryResponseBody
    type: object
    properties:
      chargeHistory:
        $ref: '#/definitions/ChargeHistoryResponseBody'
    example:
      chargeHistory:
        deviceID: Ipsa a delectus quia.
        logs:
        - charge:
            charging: true
            timestamp: 7530045597555114181
            value: 7362310245545543380
          deviceID: Quia provident rerum.
        - charge:
            charging: true
            timestamp: 7530045597555114181
            value: 7362310245545543380
          deviceID: Quia provident rerum.
        - charge:
            charging: true
            timestamp: 7530045597555114181
            value: 7362310245545543380
          deviceID: Quia provident rerum.
    required:
    - chargeHistory
  ChargewatchListDevicesResponseBody:
    title: ChargewatchListDevicesResponseBody
    type: object
    properties:
      devices:
        type: array
        items:
          $ref: '#/definitions/DeviceResponseBody'
        example:
        - charge:
            charging: false
            timestamp: 8344681023334431344
            value: 1969285450001603711
          description: Sit ea earum vel.
          id: Iusto optio.
          name: Quia quas illo sunt quae vitae.
          state: Quaerat totam.
          userID: Corporis labore accusamus sit excepturi non.
        - charge:
            charging: false
            timestamp: 8344681023334431344
            value: 1969285450001603711
          description: Sit ea earum vel.
          id: Iusto optio.
          name: Quia quas illo sunt quae vitae.
          state: Quaerat totam.
          userID: Corporis labore accusamus sit excepturi non.
        - charge:
            charging: false
            timestamp: 8344681023334431344
            value: 1969285450001603711
          description: Sit ea earum vel.
          id: Iusto optio.
          name: Quia quas illo sunt quae vitae.
          state: Quaerat totam.
          userID: Corporis labore accusamus sit excepturi non.
        - charge:
            charging: false
            timestamp: 8344681023334431344
            value: 1969285450001603711
          description: Sit ea earum vel.
          id: Iusto optio.
          name: Quia quas illo sunt quae vitae.
          state: Quaerat totam.
          userID: Corporis labore accusamus sit excepturi non.
    example:
      devices:
      - charge:
          charging: false
          timestamp: 8344681023334431344
          value: 1969285450001603711
        description: Sit ea earum vel.
        id: Iusto optio.
        name: Quia quas illo sunt quae vitae.
        state: Quaerat totam.
        userID: Corporis labore accusamus sit excepturi non.
      - charge:
          charging: false
          timestamp: 8344681023334431344
          value: 1969285450001603711
        description: Sit ea earum vel.
        id: Iusto optio.
        name: Quia quas illo sunt quae vitae.
        state: Quaerat totam.
        userID: Corporis labore accusamus sit excepturi non.
    required:
    - devices
  ChargewatchUpdateChargeResponseBody:
    title: ChargewatchUpdateChargeResponseBody
    type: object
    properties:
      device:
        $ref: '#/definitions/DeviceResponseBody'
    example:
      device:
        charge:
          charging: false
          timestamp: 8344681023334431344
          value: 1969285450001603711
        description: Sit ea earum vel.
        id: Iusto optio.
        name: Quia quas illo sunt quae vitae.
        state: Quaerat totam.
        userID: Corporis labore accusamus sit excepturi non.
    required:
    - device
  ChargewatchUpdateChargeStatusBadRequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: updateCharge_StatusBadRequest_response_body result type (default
      view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ChargewatchUpdateChargeStatusInternalServerErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: updateCharge_StatusInternalServerError_response_body result type
      (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ChargewatchUpdateDeviceResponseBody:
    title: ChargewatchUpdateDeviceResponseBody
    type: object
    properties:
      device:
        $ref: '#/definitions/DeviceResponseBody'
    example:
      device:
        charge:
          charging: false
          timestamp: 8344681023334431344
          value: 1969285450001603711
        description: Sit ea earum vel.
        id: Iusto optio.
        name: Quia quas illo sunt quae vitae.
        state: Quaerat totam.
        userID: Corporis labore accusamus sit excepturi non.
    required:
    - device
  DeviceResponseBody:
    title: 'Mediatype identifier: device; view=default'
    type: object
    properties:
      charge:
        $ref: '#/definitions/ChargeResponseBody'
      description:
        type: string
        example: Consequuntur ipsa vero assumenda dicta.
      id:
        type: string
        example: Molestiae ducimus itaque neque adipisci voluptatem.
      name:
        type: string
        example: Placeat totam error repellendus vel aut.
      state:
        type: string
        example: Perferendis totam sint expedita qui.
      userID:
        type: string
        example: Ut laudantium.
    description: DeviceResponseBody result type (default view)
    example:
      charge:
        charging: true
        timestamp: 7530045597555114181
        value: 7362310245545543380
      description: Nihil aliquam ut ea et.
      id: Illum odit sint aut quis.
      name: Saepe delectus doloribus consequatur.
      state: Ad soluta quas perspiciatis rerum ipsum asperiores.
      userID: Mollitia iste omnis.
    required:
    - id
    - userID
    - name
    - description
    - state
    - charge